+++ src/pages/dashboard/newSim.js
@@ -1,21 +1,44 @@
+import {useState} from 'react'
 import Col from "../../components/colum"
 import { deleteAttachment } from "../../store/actions/attachmentAction"
 import { useDispatch } from "react-redux"
 import { updateSimAcStatue } from "../../store/actions/addSimAction"
 import URL from '../../store/serverUrl'
+import moment from "moment";
 
  
-const NewSim = ({ attachment }) => {
+const NewSim = ({ attachment,index, length }) => {
 
    const dispatch = useDispatch()
 
+
+  const [openModel, setOpen] = useState(false)
+
+  const [modelData, setFormData] = useState({})
+
+ console.log(attachment)
    const handleDelete = () => {
+        let confirmed = window.confirm("Are you sure to delete it?")
+       if(confirmed){
         dispatch(deleteAttachment(attachment.id))
+  }
+}
+
+    
+  const handleToggleModel = () => {
+    setOpen(!openModel)
   }
   
   const handleAcitvate = () => {
-    dispatch(updateSimAcStatue(attachment.sim.id))
+    dispatch(updateSimAcStatue(attachment.sim.id,modelData))
+   setOpen(!openModel)
   }
+
+   const handleChange = e => {
+    setFormData({ ...modelData, [e.target.name]: e.target.value })
+  }
+
+
     
     return (
         <>
@@ -23,6 +46,7 @@ const NewSim = ({ attachment }) => {
              
                     <tr bgcolor="#fff1f1" className='table-raw'>
 
+                    <Col val={`${length - index}` } />
                     <Col val={attachment.sim.user.name} />
             
                     <Col val={attachment.sim.operatorName} />
@@ -30,36 +54,54 @@ const NewSim = ({ attachment }) => {
                     <Col val={attachment.sim.simNumber} />
                         
                     <Col val={attachment.sim.ICCID} />
+                    <Col val={moment(attachment.sim.soldAt).format('MMMM Do YYYY, h:mm:ss a')} />
             
-            <td className='col'><button className={attachment.sim.approvalStatus?'':'delete-button'} onClick={handleAcitvate}>{attachment.sim.approvalStatus?'Activated':'Activate'}</button></td>
+            <td className='col'><button className={attachment.sim.approvalStatus?'':'delete-button'} onClick={handleToggleModel}>{attachment.sim.approvalStatus?'Activated':'Activate'}</button></td>
                         
 
            <td className='col'>  <button className='dawnload-button'>
-                <a href={`${URL}/wordfile/${attachment.file_1}`} download > Word File </a>
+                <a   style={{ color: 'rgb(0, 39, 75)', fontWeight: '800' }} href={`${URL}/wordfile/${attachment.file_1}`} download > Word File </a>
             </button>
             </td>
             <td className='col'>
                         {
-                attachment.file_2 ? <button className='dawnload-button'> <a rel="noreferrer" target="_blank" href={`${URL}/uploads/${attachment.file_2}`} download >  File1</a></button>:''
+                attachment.file_2 ? <button className='dawnload-button'> <a   style={{ color: 'rgb(0, 39, 75)', fontWeight: '800' }} rel="noreferrer" target="_blank" href={`${URL}/uploads/${attachment.file_2}`} download >  File1</a></button>:''
               }  
               </td>
              <td className='col'> {
-                attachment.file_3 ? <button className='dawnload-button'> <a rel="noreferrer" target="_blank" href={`${URL}/uploads/${attachment.file_3}`} download >  File2</a></button>:''
+                attachment.file_3 ? <button className='dawnload-button'> <a   style={{ color: 'rgb(0, 39, 75)', fontWeight: '800' }} rel="noreferrer" target="_blank" href={`${URL}/uploads/${attachment.file_3}`} download >  File2</a></button>:''
                         }
                         </td>
                         <td className='col'>
             {
-                attachment.file_4 ? <button className='dawnload-button'> <a rel="noreferrer" target="_blank" href={`${URL}/uploads/${attachment.file_4}`} download > File3</a></button>:''
+                attachment.file_4 ? <button className='dawnload-button'> <a  style={{ color: 'rgb(0, 39, 75)', fontWeight: '800' }} rel="noreferrer" target="_blank" href={`${URL}/uploads/${attachment.file_4}`} download > File3</a></button>:''
               }  
                         </td>
-                        <td className='col'>
-            {
-                attachment.file_5 ? <button className='dawnload-button'> <a rel="noreferrer" target="_blank" href={`${URL}/uploads/${attachment.file_5}`} download > File4</a></button>:''
-              }  
-              </td>
+                        
+                     <Col val={attachment.sim.comment} />
            
               <td className='col'><button className='delete-button' onClick={handleDelete}>Delete</button></td>
             </tr>
+
+          <div className={openModel ? 'model-active' : 'model'}>
+        <div className='model-sub'>
+          <label className='mod-lab' name='comment' htmlFor='comment'>
+            Add A Comment
+          </label>
+          <input
+            onChange={handleChange}
+            className='mod-inp'
+            name='comment'
+            type='text'
+          />
+          <button className='mod-btn' onClick={handleAcitvate}>
+            Done
+          </button>
+          <button className='mod-btn-clr' onClick={handleToggleModel}>
+            Close
+          </button>
+        </div>
+      </div>
          
                
      </>
